Load WebViews with a fake base url to ensure data with special entities renders.
Without specifying a base url, all of the data the WebView will render is
included in plaintext in the data URL passed to it.

For example, when viewing a CSS file in FileViewer, the percent signs would be
included in the data URL, and the WebView would not be able to load.

Signed-off-by: Eddie Ringle <eddie@eringle.net>