Link to our new Gist Activity when we're displaying a GistEvent
This is more of a hack, really. It cheats by passing a gist ID to the MyGist
activity as well as a boolean that prevents it from actually displaying the
MyGist activity's content. It then grabs the Gist info and sends it over to
SingleGist. This should be fixed before final release of 2.0, but is good
enough for a proof of concept right now.

Signed-off-by: Eddie Ringle <eddie@eringle.net>